feat: ability to add email to recipients list by clicking enter (#760)
I encountered so many times now that I send by email, which I use 99% of the time, it didnt got send, because the mail was not accepted when pressing enter. 

Also because no confirmation mail is send, i never knew it arrived. Ill work on that code later, but adding enter to that field, will help a lot!